Override drop_table method
When a view is created on a table, it creates a dependency on that
table.  Currently, when a schema is loaded, if the table and view
already exist, the drop_table call fails, since the view still exists.

The solution is to append the `CASCADE` option on the drop_table
command.